Assignment1_9.html
Create a web page that contains two buttons. When the first is clicked, the user will be prompted to 
enter a first name, then a last name. These names will be concatenated to form a username with a dot 
between the first and last names. The second button will prompt the user for a nickname.