Create Assignment1_17.html
Write a javascript statment to accomplish each of the following tasks:
a. Assign the sum of x and y to z, and increment the value of x by 1 after the calculation. Use only 
one statment.
b. Test whether the value of the variable count is greater than 10. If it is, then print Count is 
greater than ten.
c. Decrement the variable x by 1, then subtract it from the variable total. Use only one statement.
d. Calculate the remainder after q is divided by divisor, and assign the result to q. Write this 
statment in two different ways.
e. Declare variables sum and x.
f. Assign 1 to variable x.
g. Assign 0 to variable sum.
h. Add variable x to variable sum, and assign the result to variable sum. Print The sum is: , 
followed by the value of variable sum.