Create Assignment1_18.html
Write a program that calculates an average of miles per gallon for a vehicle over the history of 
miles/gallons factor. The steps are the following:
a. Input number of full tanks to average.
b. Input the gallons of fuel filled for number x
c. Input the miles driven for number x
d. Repeat entry of gallons and miles until the number of tanks is entered.
e. Print the average miles per gallon for each tankful and print an average at the bottom.