Create Assignment1_19.html
Write a program that calculates whether a customer has exceeded the credit limit of an account. To do this:
a. Obtain account number, balance previous, total charges, total payments/credits, and credit limit.
b. Calculate current debt balance
c. Compare that the debt is < credit limit, if so print leftover credit, if not print credit exceeded