Create Assignment1_21.html
Create a program that calculates the gross pay of an employee, taking into account overtime 
compensation. To do this:
a. Take in employee name.
b. Take in employee rate pay.
c. Take in employee hours worked that week.
d. If hours over 40, then the hours minus 40 should have a rate of time and a half added to the 
total 40 * rate.
e. Print results.
f. Repeat until all employees are added.